movido a pasta de propostas para cadastros junto com area promotora
