+++ src/pages/cadastros/propostas/form/steps/formulario-detalhamento/components/carga-horaria-total/index.tsx
