+++ src/pages/cadastros/propostas/form/steps/formulario-detalhamento/formulario-detalhamento.tsx
